Updated Travis-CI Configuration (#9)
Travis-CI will use containers to run the tests now, this ensures fast execution.
The PHP versions that are tested were updated as well, nightly failures are of
course ignored. I also disabled the creation of code coverage reports as long as
they are not in use, this will be addressed in #7. Last but not least I removed
self updating of composer, just wastes time, and added multiple operating
systems to the matrix where OSX is allowed to fail due to platform instability.