Reduce amount of tests and parallelize uploads
The Mac builds are for the birds since all of them fail all the time. Better to
disable them completely and hopefully get boxes assigned faster from Travis-CI.

Also changed the approach to determine if code coverage should be produced. The
previous configuration ran the tests twice with PHP 7.1, this is not what we
wanted. Along with that I try to parallelize the uploading of the code coverage
to the various third-party websites; again to speed up the testing.