Added services for quality assurance
All services are free and it does not hurt to have many since their algorithms
to determine issues are slightly different. The fact that everything is
automated ensures that nobody has to walk through an endless amount of messages
while at the same time every interested user can have a look at the tool they
personally prefer the most and feel comfortable with.

Improved Travis-CI configuration and ensured that all services are working
fine. There is still a problem with CodeClimate but I have the feeling that
this will resolve itself as soon as some code coverage for the master branch
was submitted to them. We'll see, might require a follow up ticket. For now
this supposedly closes #7.